new namespace for book (controller & view) && book index for admin view && edit datatable_book.js to add details to book index view
