+++ app/assets/stylesheets/admin.scss
@@ -6,3 +6,4 @@
 @import "admin/fontawesome.scss.erb";
 @import "admin/jquery-ui-1.10.4.custom.css";
 @import "admin/magnific-popup.css";
+@import "admin/dataTables.bootstrap.min.css"
