+++ app/javascript/packs/admin.js
@@ -8,10 +8,10 @@ require('bootstrap')
 require('bootstrap-datepicker')
 require('placeholder')
 require('magnific-popup')
-require('datatables.net-bs4')
+require('datatables.net-bs')
 require('bootstrap-tagsinput')
 // files
-require('admin/theme')
+//require('admin/theme')
 require('admin/sidebar')
 require('admin/add_book')
 require('admin/datatable_book')
