+++ app/views/manager/books/_form.html.erb
@@ -1,4 +1,4 @@
-<%= form_with(model: book, local: true, class: "form-horizontal",
+<%= form_with(model: [ :manager, book ], local: true, class: "form-horizontal",
                 html: { autocomplete: "off" } ) do |form| %>
 <% if book.errors.any? %>
 <div id="error_explanation">
