+++ app/views/manager/books/_taginput.html.erb
