+++ app/views/manager/books/edit.html.erb
@@ -8,8 +8,8 @@
   
         <%= render layout: 'shared/admin/page_header', locals: { title: t('book.view.editing_book') } do %>
             <%= render 'form', book: @book %>
-            <%= link_to 'Show', @book %> |
-            <%= link_to 'Back', books_path %>
+            <%= link_to t('default.show'), manager_book_path(@book.id) %> |
+            <%= link_to t('default.back'), manager_books_path %>
         <% end %>
         
       </section>
