+++ app/views/manager/books/index.html.erb
@@ -1,0 +1,81 @@
+<section class="body">
+
+  <%= render partial: 'shared/admin/navbar' %>
+  <div class="inner-wrapper">
+    <%= render partial: 'shared/admin/sidebar' %>
+
+    <section role="main" class="content-body">
+
+      <%= render layout: 'shared/admin/page_header', locals: { title: t('default.book.name.other') } do %>
+        <table class="table table-bordered table-striped mb-none" id="datatable-book">
+          <thead>
+            <tr>
+              <th><%= Book.human_attribute_name(:description) %></th>
+              <th><%= t('default.image') %></th>
+              <th><%= Publisher.model_name.human %></th>
+              <th><%= t('default.author.name.other') %></th>
+              <th><%= t('default.category.name.other') %></th>
+              <th><%= Book.human_attribute_name(:title) %></th>
+              <th><%= Book.human_attribute_name(:isbn) %></th>
+              <th><%= Book.human_attribute_name(:published_date) %></th>
+              <th><%= Book.human_attribute_name(:page_count) %></th>
+              <th><%= Book.human_attribute_name(:language) %></th>
+              <th><%= Book.human_attribute_name(:is_pdf) %></th>
+              <th><%= Book.human_attribute_name(:quantity) %></th>
+              <th><%= Book.human_attribute_name(:price) %></th>
+            </tr>
+          </thead>
+          <tbody>
+            <% @books.each do |book| %>
+              <tr class="gradeX">
+                <td><%= book.description %></td>
+                <td><%= book.image_link %></td>
+                <td><%= publisher_name(book.publisher) %></td>
+                <td><%= csv_names(book.authors) %></td>
+                <td><%= csv_names(book.categories) %></td>
+                <td><%= book.title %></td>
+                <td><%= book.isbn %></td>
+                <td><%= book.published_date %></td>
+                <td><%= book.page_count %></td>
+                <td><%= book.language %></td>
+                <td><%= book.is_pdf %></td>
+                <td><%= book.quantity %></td>
+                <td><%= book.price %></td>
+                <template id="table_hidden_data">
+                  <table class="table mb-none">
+                    <div class="card mb-6">
+                      <div class="row no-gutters">
+                        <div class="col-md-2">
+                          <%= image_tag book.image_link, class: "card-img" %>
+                        </div>
+                        <div class="col-md-10">
+                          <div class="card-body">
+                            <h5 class="card-title"><%= book.title %></h5>
+                            <p class="card-text"><small class="text-muted"><% csv_names(book.authors) %></small></p>
+                            <p class="card-text"><%= book.description %></p>
+                            <p class="card-text"><small class="text-muted"><%= csv_names(book.categories) %></small></p>
+                            <p class="card-text"><small class="text-muted"><%= publisher_name(book.publisher) %></small></p>
+                            <div class="btn-group">
+                              <%= link_to t('default.show'), manager_book_path(book.id),
+                                          class: "btn btn-info" %>
+                              <%= link_to t('default.edit'), edit_manager_book_path(book.id),
+                                          class: "btn btn-primary" %> 
+                              <%= link_to t('default.destroy'), manager_book_path(book.id),
+                                          method: :delete, class: "btn btn-danger",
+                                          data: { confirm: t('default.messages.ensure') }  %>
+                            </div>
+                          </div>
+                        </div>
+                      </div>
+                    </div>
+                  </div>
+                </template>  
+              </tr>
+            <% end %>
+          </tbody>
+        </table>
+        <%= link_to t('book.view.new_book'), new_manager_book_path %>
+      <% end %>
+    </section>
+  </div>
+</section>
