+++ app/views/manager/books/new.html.erb
