+++ app/views/manager/books/show.html.erb
@@ -55,4 +55,5 @@
   <%= @book.price %>
 </p>
 
-<%= link_to 'Back', books_path %>
+<%= link_to 'Edit', edit_manager_book_path(@book) %> |
+<%= link_to 'Back', manager_books_path %>
