+++ package.json
@@ -9,8 +9,8 @@
     "bootstrap": "^3.3.0",
     "bootstrap-datepicker": "^1.9.0",
     "bootstrap-tagsinput": "^0.7.1",
-    "datatables.net": "^1.10.19",
-    "datatables.net-bs4": "^1.10.21",
+    "datatables.net": "^1.10.21",
+    "datatables.net-bs": "^1.10.21",
     "imports-loader": "^1.1.0",
     "jquery": "^3.5.1",
     "magnific-popup": "^1.1.0",
